Don't register aliased attributes in attributes_types
In rails 7.1, this was failing when trying to access the virtual
aliased attribute with: ActiveModel::MissingAttributeError.
Perhaps, we're registering it incorrectly as rails has had attribute_alias
and friends methods and supports it on it's own.

For now, we only register non-aliased attributes in attribute_types
but bring them back in virtual_attribute_names since they are already
attributes (but aliases) but still virtual.